+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `clone` command now allow to change the project and description on the
+  time entry creation, interactive mode already had this possibility
+
 ### Fixed
 
 - `clone` sub-command was not asking to confirm the tags when the original time entry already